requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
discord.py==1.7.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
moviepy==1.0.3
multidict==5.1.0
numpy==1.20.2
Pillow==8.2.0
proglog==0.1.9
python-dotenv==0.17.1
requests==2.25.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
